Re-coded to be more stable.
This is a an almost complete re-coding of the notifications script.
The previous version had too useless complexities and was computing each element position independently causing the notifications to be unstably displayed in HMD if the user moved.
There was also too many array used to achieved this. It was a bit painful.

This new version use an invisible entity parented to the avatar as a "container" where the notification elements are parented to it, making the positioning easier and more stable if the avatar moves.
I used only one array of objects to contain the the notification stack.
There is also a better management of some signal to avoid useless computing when it wasn't necessary.
I think the code is also more clear.

I got rid of the little x buttons to close the notification, it was on each element, you had to target a very small square surface with the mouse if you wanted to absolutely close a notification within the 10 second of their existence. Now you can close a notification simply by clicking on it (on all its surface)
But as before, the close function works only on desktop mode. (Apparently it's relatively complex to get the id of the entity that you click on in HMD. Maybe one day.)

It's also easier now to adjust the scale of the notifications in HMD if we want too, (only one value to change in the script)
The parenting using a container entity has also made all the notification UI in HMD scaling correctly according the avatar scale (it was free by parenting all those together).