remove Vircadia, add metax.zone
This PR remove Vircadia from the Federation (Since it is not really part of a federation when it is one way and not even working with us)
This PR also add "https://api.mv.metax.zone" as part of the federation (minimally for testing/development) 

Note: Vircadia should remain in the place app as an "external" region for those who lock it in the create app. It will be up to them to keep it or let it go.