Restore the original dimensions of the bubble.
There was a "scale" property on the original 3d overlay that was used in this script or scale the dimensions.
But it seems that the "scale" property was not doing anything. 
When I moved this script to Local Entities, I had to multiply the dimensions with their respective scale and the result was badly streched. (clearly, based on the texture deformation, it was not wished. In addition that Hifi release it as functional ignoring that the scale was doing nothing.)
This fix just remove that scaling added assuming that it was working. 
Now it is it was before V8.