Home button fallback to Tutorial
Previously, the Home button, in the case where there is no home bookmark defined, was redirecting to localhost. But some peopl emight not have a localhost domain server running. This PR change the fallback to the Tutorial instead of localhost,

In addition, the tutorial place card wasn't work correctly. (it was trying to use the name instead of the address. I added an exception case for this.