Initial effort to reorganize the startup code.
This is intended to make things like plugin initialization more sane,
and make Qt happier. Qt wants QApplication to start as soon as possible,
but our code's attempt to load the entire world in the Application
constructor doesn't quite mesh with this.

This is an effort to first create Application much earlier, and do as
little as possible in the constructor, moving almost all initialization
to a later init function.

A later stage would be to split the giant mess of init code into more
digestible sections that run in some sort of logical order.