Custom refresh rate profile WIP
This adds a Custom refresh rate profile with configurable refresh rate regimes.

Unfortunately this commit does not work yet. The SpinBoxes inside GraphicsSettings are on top of each other, and changing the values for the Custom profile through e.g. the JS API does not actually affect the refresh rate cap at all for some reason.